# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Jornada de trabalho com marcos de hidratação e exercício.",
    "title" : "Pausa Ativa API",
    "version" : "0.3.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Mesma origem do frontend"
  } ],
...
  "paths" : {
...
    "/api/v1/historico" : {
      "get" : {
        "description" : "Taxa = concluídos ÷ (concluídos + falhas), por categoria. A data não pode ser futura.",
        "operationId" : "consultar_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "periodo",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "DIA", "SEMANA", "MES" ]
          }
        }, {
          "in" : "query",
          "name" : "data",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Historico"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          }
        },
        "summary" : "Resumo do dia, da semana (segunda a domingo) ou do mês que contém a data",
        "tags" : [ "Histórico" ]
      }
    },
    "/api/v1/jornadas" : {
      "get" : {
        "description" : "No mesmo formato de /jornadas/atual. Usada pelo histórico para rever um dia passado.",
        "operationId" : "doDia",
        "parameters" : [ {
          "in" : "query",
          "name" : "data",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "Houve jornada no dia"
          },
          "204" : {
            "description" : "Nenhuma jornada no dia"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Jornada de um dia, com os blocos propostos",
        "tags" : [ "Jornada" ]
      },
      "post" : {
        "description" : "Exige o perfil físico preenchido: sem ele, responde 409.",
        "operationId" : "iniciar",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IniciarJornada"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Inicia o dia. Encerra antes a jornada esquecida de um dia anterior, se houver.",
        "tags" : [ "Jornada" ]
      }
    },
    "/api/v1/jornadas/atual" : {
      "get" : {
        "operationId" : "atual",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "Há jornada aberta ou de hoje"
          },
          "204" : {
            "description" : "Nenhuma jornada hoje"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Jornada aberta ou, sem ela, a de hoje",
        "tags" : [ "Jornada" ]
      }
    },
    "/api/v1/jornadas/{id}/finalizacao" : {
      "post" : {
        "operationId" : "finalizar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Finaliza o dia. Repetir devolve a mesma situação.",
        "tags" : [ "Jornada" ]
      }
    },
    "/api/v1/jornadas/{id}/pausa" : {
      "post" : {
        "operationId" : "pausar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Pausa a jornada (almoço). O tempo trabalhado congela.",
        "tags" : [ "Jornada" ]
      }
    },
    "/api/v1/jornadas/{id}/retomada" : {
      "post" : {
        "operationId" : "retomar",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Jornada"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Conflict"
          }
        },
        "summary" : "Retoma a jornada pausada",
        "tags" : [ "Jornada" ]
      }
    },
...
    "/api/v1/perfil" : {
      "get" : {
        "operationId" : "consultar",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Perfil"
                }
              }
            },
            "description" : "Perfil preenchido"
          },
          "204" : {
            "description" : "Perfil ainda não preenchido"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          }
        },
        "summary" : "Perfil físico atual",
        "tags" : [ "Perfil" ]
      },
      "put" : {
        "operationId" : "salvar",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SalvarPerfil"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Perfil"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Bad Request"
          }
        },
        "summary" : "Cria ou substitui o perfil físico. Vale para os próximos blocos.",
        "tags" : [ "Perfil" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Bloco" : {
        "type" : "object",
        "description" : "Bloco de exercício do marco, como foi proposto no disparo.",
        "properties" : {
          "compensaAdiamento" : {
            "type" : "boolean",
            "description" : "Se o bloco compensa o anterior, que foi adiado; nesse caso tem 10 min"
          },
          "duracaoMin" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Duração do bloco: 5 ou 10 min",
            "example" : 5
          },
          "itens" : {
            "type" : "array",
            "description" : "Exercícios na ordem em que devem ser feitos",
            "items" : {
              "$ref" : "#/components/schemas/ItemDoBloco"
            }
          },
          "segundosEstimados" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Soma das estimativas dos exercícios, com as trocas",
            "example" : 278
          }
        },
        "required" : [ "compensaAdiamento", "duracaoMin", "itens", "segundosEstimados" ]
      },
      "ContagemDoDia" : {
        "type" : "object",
        "description" : "Uma categoria num dia: situações, taxa e meta.",
        "properties" : {
          "categoria" : {
            "type" : "string",
            "enum" : [ "HIDRATACAO", "EXERCICIO" ]
          },
          "concluidos" : {
            "type" : "integer",
            "format" : "int32"
          },
          "emAberto" : {
            "type" : "integer",
            "format" : "int32"
          },
          "falhas" : {
            "type" : "integer",
            "format" : "int32"
          },
          "metaAtingida" : {
            "type" : [ "boolean", "null" ],
            "description" : "Se o dia chegou a 80%. Nula sem dados."
          },
          "naoConcluidos" : {
            "type" : "integer",
            "format" : "int32"
          },
          "naoEntregues" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taxa" : {
            "type" : [ "number", "null" ],
            "description" : "Taxa do dia, em %, truncada em uma casa. Nula sem dados.",
            "example" : 85.7
          }
        },
        "required" : [ "categoria", "concluidos", "emAberto", "falhas", "metaAtingida", "naoConcluidos", "naoEntregues", "taxa" ]
      },
...
      "DiaDoHistorico" : {
        "type" : "object",
        "description" : "Um dia do período. Sem jornada, não tem categorias e não conta como falha.",
        "properties" : {
          "categorias" : {
            "type" : "array",
            "description" : "Água e exercício, nessa ordem; vazia sem jornada",
            "items" : {
              "$ref" : "#/components/schemas/ContagemDoDia"
            }
          },
          "data" : {
            "type" : "string",
            "format" : "date"
          },
          "futuro" : {
            "type" : "boolean",
            "description" : "Dia depois de hoje, dentro da semana ou do mês corrente"
          },
          "jornada" : {
            "type" : [ "string", "null" ],
            "description" : "Estado da jornada do dia. Nulo sem jornada.",
            "enum" : [ "EM_ANDAMENTO", "PAUSADA", "FINALIZADA", "ENCERRADA_AUTOMATICAMENTE" ]
          }
        },
        "required" : [ "categorias", "data", "futuro", "jornada" ]
      },
      "Historico" : {
        "type" : "object",
        "description" : "Uma visão do histórico: o resumo de cada categoria e todos os dias do período.",
        "properties" : {
          "categorias" : {
            "type" : "array",
            "description" : "Água e exercício, nessa ordem",
            "items" : {
              "$ref" : "#/components/schemas/ResumoDaCategoria"
            }
          },
          "dias" : {
            "type" : "array",
            "description" : "Todos os dias do período, em ordem, com ou sem jornada",
            "items" : {
              "$ref" : "#/components/schemas/DiaDoHistorico"
            }
          },
          "fim" : {
            "type" : "string",
            "format" : "date",
            "description" : "Último dia do período, inclusive"
          },
          "inicio" : {
            "type" : "string",
            "format" : "date",
            "description" : "Primeiro dia do período"
          },
          "periodo" : {
            "type" : "string",
            "enum" : [ "DIA", "SEMANA", "MES" ]
          }
        },
        "required" : [ "categorias", "dias", "fim", "inicio", "periodo" ]
      },
      "IniciarJornada" : {
        "type" : "object",
        "properties" : {
          "duracaoBlocoMin" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 5,
            "description" : "Duração dos blocos de exercício do dia, em minutos: 5 ou 10. Sem ela, 5 min.",
            "example" : 5
          },
          "metaAguaMl" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 3000,
            "description" : "Meta de água do dia, em ml. Sem ela, 3.000 ml.",
            "example" : 3000,
            "maximum" : 6000,
            "minimum" : 1
          }
        }
      },
      "ItemDoBloco" : {
        "type" : "object",
        "properties" : {
          "exercicio" : {
            "type" : "string",
            "example" : "Sentar e levantar da cadeira"
          },
          "grupo" : {
            "type" : "string",
            "example" : "Pernas"
          },
          "instrucao" : {
            "type" : "string",
            "example" : "Sente e levante da cadeira sem usar as mãos, com os pés na largura do quadril."
          },
          "quantidade" : {
            "type" : "string",
            "description" : "\"10 repetições\", \"6 por lado\" ou \"20 s\"",
            "example" : "10 repetições"
          }
        },
        "required" : [ "exercicio", "grupo", "instrucao", "quantidade" ]
      },
      "Jornada" : {
        "type" : "object",
        "description" : "Situação da jornada no instante calculadoEm. Horários no fuso America/Sao_Paulo.",
        "properties" : {
          "aguaIngeridaMl" : {
            "type" : "number",
            "description" : "Soma dos marcos de hidratação concluídos",
            "example" : 562.5
          },
          "calculadoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dataReferencia" : {
            "type" : "string",
            "format" : "date",
            "description" : "Dia em que a jornada começou"
          },
          "duracaoBlocoMin" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Duração dos blocos de exercício escolhida ao iniciar o dia. O bloco que compensa um adiamento tem 10 min.",
            "example" : 5
          },
          "finalizadaEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "iniciadaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "marcos" : {
            "type" : "array",
            "description" : "Água e exercício em ordem de horário; na hora cheia, a água vem antes",
            "items" : {
              "$ref" : "#/components/schemas/Marco"
            }
          },
          "metaAguaMl" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 3000
          },
          "pausadaDesde" : {
            "type" : [ "string", "null" ],
            "format" : "date-time",
            "description" : "Início da pausa em curso; nulo fora da pausa"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "EM_ANDAMENTO", "PAUSADA", "FINALIZADA", "ENCERRADA_AUTOMATICAMENTE" ]
          },
          "tempoTrabalhadoSegundos" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Tempo trabalhado até calculadoEm, sem as pausas. Enquanto a jornada estiver em andamento, a tela soma o tempo decorrido desde então.",
            "example" : 5400
          }
        },
        "required" : [ "aguaIngeridaMl", "calculadoEm", "dataReferencia", "duracaoBlocoMin", "finalizadaEm", "id", "iniciadaEm", "marcos", "metaAguaMl", "pausadaDesde", "status", "tempoTrabalhadoSegundos" ]
      },
      "Marco" : {
        "type" : "object",
        "description" : "Lembrete da jornada. Horários no fuso America/Sao_Paulo.",
        "properties" : {
          "bloco" : {
            "description" : "Bloco de exercício. Nulo na água e no exercício que ainda não disparou.",
            "oneOf" : [ {
              "$ref" : "#/components/schemas/Bloco"
            }, {
              "type" : "null"
            } ]
          },
          "categoria" : {
            "type" : "string",
            "enum" : [ "HIDRATACAO", "EXERCICIO" ]
          },
          "disparadoEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "editadoEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time",
            "description" : "Instante da última correção. Nulo se o lembrete nunca foi corrigido."
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "mensagem" : {
            "type" : "string",
            "example" : "Beba ~190 ml. Levante-se para buscar a água."
          },
          "podeAdiar" : {
            "type" : "boolean",
            "description" : "Se o botão Adiar vale agora: só no exercício pendente que não compensa um adiamento e não é o último do dia"
          },
          "podeCorrigir" : {
            "type" : "boolean",
            "description" : "Se o botão Corrigir vale agora: só no lembrete concluído ou com falha, até o fim do dia em que a jornada começou"
          },
          "recebidoEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "respondidoEm" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "segundosTrabalhadosPrevistos" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Tempo trabalhado em que o marco dispara",
            "example" : 1800
          },
          "sequencia" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Ordem do marco na categoria, a partir de 1",
            "example" : 1
          },
          "status" : {
            "type" : "string",
            "enum" : [ "AGENDADO", "PENDENTE", "CONCLUIDO", "FALHA", "NAO_ENTREGUE", "NAO_CONCLUIDO", "ADIADO" ]
          },
          "volumeAproximadoMl" : {
            "type" : [ "integer", "null" ],
            "format" : "int32",
            "description" : "Volume para exibir, arredondado para a dezena. Nulo no exercício.",
            "example" : 190
          },
          "volumeMl" : {
            "type" : [ "number", "null" ],
            "description" : "Volume exato: meta ÷ 16. Nulo no exercício.",
            "example" : 187.5
          }
        },
        "required" : [ "bloco", "categoria", "disparadoEm", "editadoEm", "id", "mensagem", "podeAdiar", "podeCorrigir", "recebidoEm", "respondidoEm", "segundosTrabalhadosPrevistos", "sequencia", "status", "volumeAproximadoMl", "volumeMl" ]
      },
      "Perfil" : {
        "type" : "object",
        "description" : "Perfil físico do usuário. Vale para os blocos montados depois da última alteração.",
        "properties" : {
          "aceitaChao" : {
            "type" : "boolean"
          },
          "articulacoesPoupadas" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "JOELHO", "OMBRO", "PUNHO", "LOMBAR", "CERVICAL" ]
            }
          },
          "equipamentos" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "APOIO_DE_FLEXAO", "HALTERES_2KG" ]
            }
          },
          "nivel" : {
            "type" : "string",
            "enum" : [ "INICIANTE", "INTERMEDIARIO" ]
          }
        },
        "required" : [ "aceitaChao", "articulacoesPoupadas", "equipamentos", "nivel" ]
      },
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri"
          },
          "properties" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "format" : "uri"
          }
        }
      },
      "ResumoDaCategoria" : {
        "type" : "object",
        "description" : "Uma categoria no período: situações somadas, taxa do período e dias na meta.",
        "properties" : {
          "categoria" : {
            "type" : "string",
            "enum" : [ "HIDRATACAO", "EXERCICIO" ]
          },
          "concluidos" : {
            "type" : "integer",
            "format" : "int32"
          },
          "diasComDados" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Dias com pelo menos um concluído ou uma falha na categoria",
            "example" : 4
          },
          "diasNaMeta" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Desses dias, os que tiveram taxa de pelo menos 80%",
            "example" : 3
          },
          "emAberto" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Agendados, pendentes e adiados sem destino, só na jornada em andamento"
          },
          "falhas" : {
            "type" : "integer",
            "format" : "int32"
          },
          "metaAtingida" : {
            "type" : [ "boolean", "null" ],
            "description" : "Se a taxa chegou a 80%, pela fração exata. Nula sem dados."
          },
          "naoConcluidos" : {
            "type" : "integer",
            "format" : "int32"
          },
          "naoEntregues" : {
            "type" : "integer",
            "format" : "int32"
          },
          "taxa" : {
            "type" : [ "number", "null" ],
            "description" : "Concluídos ÷ (concluídos + falhas) do período, em %, com uma casa, truncado. Nula sem nenhum concluído nem falha.",
            "example" : 85.7
          }
        },
        "required" : [ "categoria", "concluidos", "diasComDados", "diasNaMeta", "emAberto", "falhas", "metaAtingida", "naoConcluidos", "naoEntregues", "taxa" ]
      },
      "SalvarPerfil" : {
        "type" : "object",
        "properties" : {
          "aceitaChao" : {
            "type" : "boolean",
            "description" : "Se exercícios deitado ou de quatro podem entrar no bloco"
          },
          "articulacoesPoupadas" : {
            "type" : "array",
            "description" : "Articulações a poupar. Vazio ou ausente: nenhuma.",
            "items" : {
              "type" : "string",
              "enum" : [ "JOELHO", "OMBRO", "PUNHO", "LOMBAR", "CERVICAL" ]
            }
          },
          "equipamentos" : {
            "type" : "array",
            "description" : "Equipamentos disponíveis. Cadeira e mesa já contam como disponíveis.",
            "items" : {
              "type" : "string",
              "enum" : [ "APOIO_DE_FLEXAO", "HALTERES_2KG" ]
            }
          },
          "nivel" : {
            "type" : "string",
            "enum" : [ "INICIANTE", "INTERMEDIARIO" ]
          }
        },
        "required" : [ "aceitaChao", "nivel" ]
      },
...
    }
  }
}